+++ castle/cms/static/patterns/castledynamicform.js
@@ -2,7 +2,7 @@
 
 define([
   'jquery',
-  'mockup-patterns-base',
+  'pat-base',
   'mockup-utils'
 ], function($, Base, utils) {
   'use strict';