+++ castle/cms/static/patterns/previewselect.js
@@ -2,7 +2,7 @@
 
 define([
   'jquery',
-  'mockup-patterns-base',
+  'pat-base',
   'underscore'
 ], function($, Base, _) {
   'use strict';